+++ packages/kirby/novelties/database/seeds/DefaultNoveltyTypesSeed.php
@@ -150,6 +150,24 @@ class DefaultNoveltyTypesSeed extends Seeder
             'apply_on_time_slots' => null, // any time
             'operator' => NoveltyTypeOperator::Subtraction,
         ],
+        [
+            'code' => 'PPNR',
+            'name' => 'Permiso personal NO remunerado',
+            'context_type' => null,
+            'time_zone' => 'America/Bogota',
+            'apply_on_days_of_type' => null, // any day
+            'apply_on_time_slots' => null, // any time
+            'operator' => NoveltyTypeOperator::Subtraction,
+        ],
+        [
+            'code' => 'PPR',
+            'name' => 'Permiso personal remunerado',
+            'context_type' => null,
+            'time_zone' => 'America/Bogota',
+            'apply_on_days_of_type' => null, // any day
+            'apply_on_time_slots' => null, // any time
+            'operator' => NoveltyTypeOperator::Addition,
+        ],
     ];
 
     /**
@@ -160,7 +178,7 @@ class DefaultNoveltyTypesSeed extends Seeder
         collect($this->noveltyTypes)->map(function (array $noveltyType) {
             $keys = Arr::only($noveltyType, ['code']);
 
-            return NoveltyType::updateOrCreate($keys, $noveltyType);
+            return NoveltyType::firstOrCreate($keys, $noveltyType);
         });
     }
 }