+++ packages/kirby/novelties/src/Novelties.php
@@ -4,6 +4,7 @@ namespace Kirby\Novelties;
 
 use Illuminate\Support\Collection;
 use Kirby\Novelties\Contracts\NoveltyTypeRepositoryInterface;
+use Kirby\Novelties\Models\NoveltyType;
 use Kirby\TimeClock\Contracts\SettingRepositoryInterface;
 
 /**
@@ -32,12 +33,47 @@ class Novelties
         });
     }
 
+    public function defaultAdditionNoveltyTypeId(): int
+    {
+        return $this->rawSettings()->firstWhere('key', 'novelties.default-addition-novelty-type')->value;
+    }
+
     public function defaultSubTractNoveltyTypeId(): int
+    {
+        return $this->rawSettings()->firstWhere('key', 'novelties.default-subtraction-novelty-type')->value;
+    }
+
+    public function defaultSubTractNoveltyType(): NoveltyType
+    {
+        return app(NoveltyTypeRepositoryInterface::class)->find($this->defaultSubTractNoveltyTypeId());
+    }
+
+    public function defaultAdditionNoveltyType(): NoveltyType
+    {
+        return app(NoveltyTypeRepositoryInterface::class)->find($this->defaultAdditionNoveltyTypeId());
+    }
+
+    public function isDefaultForSubtraction(NoveltyType $noveltyType): bool
+    {
+        return $noveltyType->id === $this->defaultSubTractNoveltyTypeId();
+    }
+
+    public function isDefaultForAddition(NoveltyType $noveltyType): bool
+    {
+        return $noveltyType->id === $this->defaultAdditionNoveltyTypeId();
+    }
+
+    public function isDefaultForAdditionOrSubtraction(NoveltyType $noveltyType): bool
+    {
+        return $this->isDefaultForAddition($noveltyType) || $this->isDefaultForSubtraction($noveltyType);
+    }
+
+    public function defaultSubTractBalanceNoveltyTypeId(): int
     {
         return $this->rawSettings()->firstWhere('key', 'novelties.default-subtraction-balance-novelty-type')->value;
     }
 
-    public function defaultAdditionNoveltyTypeId(): int
+    public function defaultAdditionBalanceNoveltyTypeId(): int
     {
         return $this->rawSettings()->firstWhere('key', 'novelties.default-addition-balance-novelty-type')->value;
     }