+++ packages/kirby/novelties/tests/integration/Actions/CreateManyNoveltiesActionTest.php
@@ -50,6 +50,7 @@ class CreateManyNoveltiesActionTest extends \Tests\TestCase
 
     /**
      * @test
+     *
      * @dataProvider successCases
      *
      * @param  mixed  $employees