+++ packages/kirby/time-clock/src/Actions/LogCheckIn.php
@@ -10,6 +10,7 @@ use Kirby\Employees\Models\Identification;
 use Kirby\Novelties\Contracts\NoveltyRepositoryInterface;
 use Kirby\Novelties\Contracts\NoveltyTypeRepositoryInterface;
 use Kirby\Novelties\Enums\NoveltyTypeOperator;
+use Kirby\Novelties\Facades\Novelties;
 use Kirby\TimeClock\Contracts\SettingRepositoryInterface;
 use Kirby\TimeClock\Contracts\TimeClockLogRepositoryInterface;
 use Kirby\TimeClock\Exceptions\AlreadyCheckedInException;
@@ -176,11 +177,11 @@ class LogCheckIn
         }
 
         if ($isTooLate && ! $noveltyTypeId && ! $noveltyTypeIsRequired) {
-            $noveltyType = $this->noveltyTypeRepository->findDefaultForSubtraction();
+            $noveltyType = Novelties::defaultSubTractNoveltyType();
         }
 
         if ($isTooEarly && ! $noveltyTypeId && ! $noveltyTypeIsRequired) {
-            $noveltyType = $this->noveltyTypeRepository->findDefaultForAddition();
+            $noveltyType = Novelties::defaultAdditionNoveltyType();
         }
 
         $timeClockLog = [